+++ src/Kafka/Consumer.cpp
@@ -87,7 +87,7 @@ void Consumer::addTopic(std::string const &Topic) {
 
 std::pair<PollStatus, FileWriter::Msg> Consumer::poll() {
   auto KafkaMsg = std::unique_ptr<RdKafka::Message>(
-      KafkaConsumer->consume(ConsumerBrokerSettings.PollTimeoutMS));
+      KafkaConsumer->consume(toMilliSeconds(ConsumerBrokerSettings.PollTimeout)));
   switch (KafkaMsg->err()) {
   case RdKafka::ERR_NO_ERROR: {
     auto MetaData = FileWriter::MessageMetaData{