+++ src/Stream/Partition.cpp
@@ -110,8 +110,8 @@ bool Partition::shouldStopBasedOnPollStatus(Kafka::PollStatus CStatus) {
       break;
     case PartitionFilter::StopReason::TIMEOUT:
       LOG_ERROR("Stopping consumption of data from Kafka in partition {} of "
-                "topic {} due to timeout when polling for new data.",
-                PartitionID, Topic);
+                "topic {} due to timeout ({:.1f}s passed) when polling for new data.",
+                PartitionID, Topic, double(std::chrono::duration_cast<std::chrono::milliseconds>(system_clock::now() - StopTester.getErrorTime()).count()) / 1000.0);
       break;
     case PartitionFilter::StopReason::END_OF_PARTITION:
       LOG_INFO(